Backend: Reduce availableTickets when successfully booked, refactoring

+++ backend/event-service/src/main/java/lk/nibm/eventservice/repository/EventRepository.java
@@ -2,13 +2,12 @@ package lk.nibm.eventservice.repository;
 
 import lk.nibm.eventservice.model.Event;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
 @Repository
-public interface EventRepository extends JpaRepository<Event, Integer> {
+public interface EventRepository extends JpaRepository<Event, Long> {
 
     /**
      * Find events by title with case-insensitive matching.
